Make TaskFailedException include error details by default
Currently, if a Pulp task fails and propagates upwards to generic
code which catches and logs all exceptions, the logs would show
something like this level of detail:

    Task 91e3fc2d-de01-4ddd-a236-b2179d3bba0f failed

More details are available, including the Pulp traceback, but
you'd have to explicitly catch TaskFailedException and look at
task.error_details to get at them. In practice nobody is doing that,
so the details end up missing from logs.

It makes more sense to include all the details by default, so let's
start doing that. An exception to this is the case of a retry; we'll
keep the logs terse in that case, until all retries are exhausted.