Fix upload_comps_xml omitting false 'default' fields [RHELDST-10597]
The logic here was to include a 'default' field within options in the
generated units only if the value was True.

That causes a crash during publish in Pulp, because the Pulp code at [1]
unconditionally assumes that every option has a 'default' field and will
raise a KeyError otherwise. Fix it to be compatible with Pulp's
expectation.

This is not a new bug, but was recently found as this code has been
tested with more realistic data.

[1] https://github.com/pulp/pulp_rpm/blob/pulp-rpm-2.14.3-1/plugins/pulp_rpm/plugins/distributors/yum/metadata/package.py#L162