Fix fake client's handling of deps with no version
When the fake client generated RpmDependency objects during upload, it
assumed that every dependency found in the header would have an EVR.
In fact this is not true: it's common for RPMs to declare many
dependencies using only a name and with no associated EVR. It's also
possible to declare a dependency with a version only, and no release.

Previously attempting to upload such an RPM with the fake client would
crash in parse_evr. Fix it and add a test which covers that case.
Based on existing data in Pulp, the default values in that case
should apparently be None rather than empty strings.