Make ModulemdUnit and ModulemdDefaultUnit profiles immutable
The profiles value for ModulemdUnit and ModulemdDefaultUnit was a dict
type, which ment the value was mutable. This made it dangerous to use
these classes in threads and to hash them. This change replaces the dict
type with frozendict, which is an implementation of an immutable
dictionary. Frozendict is only implemented for python3 so naturally
the change will only be available in python3. We assume that people
using this library will test their code against python3 so it will
catch any cases where they try to modify a value in profiles.