Support limiting returned fields during search [RHELDST-11485]
The Pulp API supports returning a subset of fields during search, but
pulplib API didn't support this. Add it now. When constructing a
Criteria for a particular unit type, it is possible to specify the
desired fields associated with that type.

This was previously not added in the interest of keeping the API
simpler, but according to continued analysis of pubtools-pulp-push
performance it may be needed to reduce the memory usage.

That command currently discards many unnecessary fields immediately
after search, but that approach still consumes a lot of memory over
time. This is likely due to heap fragmentation as we keep allocating
and discarding memory for the unused fields. It is better to never
construct objects for the unneeded fields in the first place.