Fix flaky search on 'version' field
FileUnit.version and ModulemdUnit.version are two fields both named
"version", but stored in Pulp in a different way, one of them being
top-level and the other stored under pulp_user_metadata.

Previously, the code responsible for generating the pulp search did not
take into account exactly what kind of unit is being searched for, and
it would end up picking one of the two Pulp fields essentially at
random. Fix it to first calculate which unit type is being queried and
then use that to resolve the ambiguity.

In cases where criteria remains ambiguous, an error is now raised on
search rather than picking one of the fields at random. It could
technically be considered a backwards-incompatible change, however any
clients currently using such a search might be getting incorrect results
at random without realizing it, so it seems best to introduce the check
even if it might break some existing code.