Adding repo locks using notes field [RHELDST-14968]
There are some opperations in pulp which should not be done in parallel.
Multiple instances of the same task can cause a race condition and
overwrite data.

This change uses a repo's notes field to implement a locking system.
This signals to other tasks that the repo is being worked on and should
not be modified until the lock is removed.