Add removal function for ud_file_release_mappings_2 [RHELDST-16431]
Before this MR it was possible to only add new mappings to the
ud_file_mappings_2 repo note. This meant that if a file was
removed from a Pulp repo, the mapping remained in the note.
This MR introduces functionality to remove mappings for files
that are no longer in that repo.