Don't overwrite headers passed to Client object
requests module populates Session object with several default
headers. The behavior in pubtools-pulpib would cause the default headers
dictionary to be completely overwritten by the one passed in kwargs.
This is cumbersome in cases where a developer wants to set only a single
header, e.g. `User-Agent`. The new behavior ensures the default values
are retained and the new values passed in kwargs only update
the existing headers dictionary.