Reimplement naturalsize as part of pubtools-pulplib [RHELDST-29440]
This commit pulls out the `naturalsize()` function from humanize module
and reimplements its equivalent in pubtools-pulplib. The humanize module
is then removed from dependencies as it's not used anywhere else.
This is done to reduce the dependency bloat and make the RPM packaging
easier on RHEL 9 which doesn't ship python-humanize.