Add filelist and get_files() to RpmUnit [RHELDST-34141]
This commit adds a `filelist` attr to RpmUnit, which maps to
file paths from Pulp `repodata.filelists` field.
Moreover, it adds a `get_files()` method to RpmUnit, which
returns combined data from this new `filelist`and `files` attrs.

This is to be used in ubi-manifest, to avoid reporting false positive
file depsolving warnings, because there are cases in which
the `files.file` field is missing.